receipt/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-documentai==2.31.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.44.0
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.0
pydantic_core==2.23.2
pytest==8.3.2
python-dotenv==1.0.1
rapidfuzz==3.9.7
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
zipp==3.20.1
